refactor utils/env.py
