+++ conan/tools/qbs/__init__.py
@@ -1,4 +1,2 @@
-from conan.tools.qbs.qbsprofile import QbsProfile
-from conan.tools.qbs.qbsprofile import QbsProfile as QbsToolchain
 from conan.tools.qbs.qbs import Qbs
 from conan.tools.qbs.qbsdeps import QbsDeps
