+++ conans/server/server_launcher.py
@@ -1,8 +1,9 @@
+import os
+
 from conans.server.launcher import ServerLauncher
 
-from conans.util.env import get_env
 
-launcher = ServerLauncher(server_dir=get_env("CONAN_SERVER_HOME"))
+launcher = ServerLauncher(server_dir=os.environ.get("CONAN_SERVER_HOME"))
 app = launcher.server.root_app
