+++ test/functional/settings/cppstd/compiler_cppstd_test.py
@@ -4,7 +4,7 @@ import unittest
 
 from parameterized.parameterized import parameterized_class
 
-from conans.util.env import environment_update
+from conan.test.utils.env import environment_update
 from conan.test.utils.test_files import temp_folder
 from conan.test.utils.tools import TestClient
 from conans.util.files import save
