+++ test/integration/command/config_test.py
@@ -6,7 +6,7 @@ from conan.api.conan_api import ConanAPI
 from conans.model.conf import BUILT_IN_CONFS
 from conan.test.utils.test_files import temp_folder
 from conan.test.utils.tools import TestClient
-from conans.util.env import environment_update
+from conan.test.utils.env import environment_update
 
 
 def test_missing_subarguments():
