+++ test/integration/command/upload/test_upload_patterns.py
@@ -4,7 +4,7 @@ import pytest
 
 from conan.test.assets.genconanfile import GenConanfile
 from conan.test.utils.tools import TestClient
-from conans.util.env import environment_update
+from conan.test.utils.env import environment_update
 
 
 class TestUploadPatterns:
