+++ test/integration/command_v2/custom_commands_test.py
@@ -7,7 +7,7 @@ import pytest
 from conan.test.assets.genconanfile import GenConanfile
 from conan.test.utils.test_files import temp_folder
 from conan.test.utils.tools import TestClient
-from conans.util.env import environment_update
+from conan.test.utils.env import environment_update
 
 
 class TestCustomCommands:
