+++ test/integration/command_v2/test_output.py
@@ -1,6 +1,6 @@
 from conan.test.assets.genconanfile import GenConanfile
 from conan.test.utils.tools import TestClient
-from conans.util.env import environment_update
+from conan.test.utils.env import environment_update
 
 
 class TestOutputLevel:
