+++ test/integration/configuration/test_profile_jinja.py
@@ -5,7 +5,7 @@ import os
 from conan import conan_version
 from conan.test.assets.genconanfile import GenConanfile
 from conan.test.utils.tools import TestClient
-from conans.util.env import environment_update
+from conan.test.utils.env import environment_update
 
 
 def test_profile_template():
