+++ test/integration/remote/auth_test.py
@@ -14,7 +14,7 @@ from conan.test.utils.test_files import temp_folder
 from conan.test.utils.tools import TestClient
 from conan.test.utils.tools import TestRequester
 from conan.test.utils.tools import TestServer
-from conans.util.env import environment_update
+from conan.test.utils.env import environment_update
 from conans.util.files import save
 
 conan_content = """
